refactor: improve performance
