+++ packages/dapp/src/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useUserTracking/';
-export { useWidgetConfig } from './useWidgetConfig';
